% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fire.R
\docType{data}
\name{fire}
\alias{fire}
\title{Fire data set
A complete data set of measurements of 15 fires with their respective distances and damage}
\format{
A data frame

 @source not applicable
}
\usage{
fire
}
\description{
Fire data set
A complete data set of measurements of 15 fires with their respective distances and damage
}
\keyword{datasets}
