% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/my_getwd.R
\name{my_getwd}
\alias{my_getwd}
\title{Get current working directory}
\usage{
my_getwd()
}
\value{
A character string representing the current working directory path
}
\description{
Get current working directory
}
\examples{
my_getwd()
}
