% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{Calculate confidence interval for a sample mean}
\usage{
myci(x)
}
\arguments{
\item{x}{Numeric vector of data}
}
\value{
Named vector containing lower and upper bounds of 95% confidence interval
}
\description{
Calculate confidence interval for a sample mean
}
\examples{
data <- rnorm(30)
myci(data)
}
