% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{Plot normal distribution with shaded area}
\usage{
myncurve(mu, sigma, a)
}
\arguments{
\item{mu}{Mean of the normal distribution}

\item{sigma}{Standard deviation of the normal distribution}

\item{a}{Point up to which to calculate and shade the cumulative probability}
}
\value{
A list containing:
  \item{mu}{Input mean}
  \item{sigma}{Input standard deviation}
  \item{a}{Input point}
  \item{area}{Cumulative probability up to point a}
}
\description{
Plot normal distribution with shaded area
}
\examples{
myncurve(0, 1, 1.96)
myncurve(100, 15, 120)
}
