% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myNRML.R
\name{myNRML}
\alias{myNRML}
\title{Newton-Raphson Maximum Likelihood Estimation with Visualization}
\usage{
myNRML(x0, delta = 0.001, llik, xrange, parameter = "param")
}
\arguments{
\item{x0}{Numeric. The initial guess for the parameter value to start the Newton-Raphson iterations.}

\item{delta}{Numeric. A small step size used for numerical approximation of derivatives. Default is 0.001.}

\item{llik}{Function. A log-likelihood function defined by the user, which takes the parameter as an argument
and returns the log-likelihood value.}

\item{xrange}{Numeric vector. A range of values (c(min, max)) over which the log-likelihood and its
derivative are plotted.}

\item{parameter}{Character. The name of the parameter being estimated. Used for axis labels in the plots.}
}
\value{
A list containing:
\item{x}{Numeric vector. The sequence of parameter estimates generated by the Newton-Raphson algorithm.}
\item{y}{Numeric vector. The values of the derivative of the log-likelihood function at each estimate.}
}
\description{
Newton-Raphson Maximum Likelihood Estimation with Visualization
}
\examples{
# Define a log-likelihood function (e.g., for a normal distribution)
ll <- function(mu) sum(dnorm(c(1, 2, 3), mean = mu, sd = 1, log = TRUE))

# Run the Newton-Raphson algorithm with visualization
myNRML(0, delta = 0.001, llik = ll, xrange = c(-1, 4), parameter = "mu")
}
