% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mysquare.R
\name{mysquare}
\alias{mysquare}
\title{Square a Number}
\usage{
mysquare(x)
}
\arguments{
\item{x}{Numeric. A value or vector of values to be squared.}
}
\value{
Numeric. The square of the input value(s).
}
\description{
This function computes the square of a given numeric input.
}
\examples{
# Square a single number
mysquare(4) # Returns 16
}
