% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_fire.R
\name{plot_fire}
\alias{plot_fire}
\title{Plot fire damage vs distance with fitted curve}
\usage{
plot_fire(x, y, title = "Fire Damage vs Distance")
}
\arguments{
\item{x}{Vector of distance values}

\item{y}{Vector of damage values}

\item{title}{Optional plot title}
}
\value{
A plot of the fire damage data with fitted line
}
\description{
Plot fire damage vs distance with fitted curve
}
\examples{
data(fire)
plot_fire(fire$DISTANCE, fire$DAMAGE)
}
