% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_scores.R
\name{z_scores}
\alias{z_scores}
\title{Calculate Z-scores and their summary statistics}
\usage{
z_scores(x)
}
\arguments{
\item{x}{A numeric vector of values to standardize}
}
\value{
A list containing:
  \item{mean_z}{The mean of the z-scores, rounded to 4 decimal places}
  \item{var_z}{The variance of the z-scores}
}
\description{
Calculate Z-scores and their summary statistics
}
\examples{
data <- c(2, 4, 6, 8, 10)
z_scores(data)
}
